feat: 키워드가 포함된 문제집 목록 검색

+++ backend/src/main/java/com/ssafy/domain/quiz/repository/QuestionRepository.java
@@ -3,5 +3,5 @@ package com.ssafy.domain.quiz.repository;
 import com.ssafy.domain.quiz.entity.Question;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface QuestionRepository extends JpaRepository<Question, Long> {
+public interface QuestionRepository extends JpaRepository<Question, Integer> {
 }
